Refactor: Move the reg_key class to a separate file
