+++ src/windows/webview_msedge.h
@@ -203,67 +203,6 @@ struct shcore_symbols {
       library_symbol<SetProcessDpiAwareness_t>("SetProcessDpiAwareness");
 };
 
-class reg_key {
-public:
-  explicit reg_key(HKEY root_key, const wchar_t *sub_key, DWORD options,
-                   REGSAM sam_desired) {
-    HKEY handle;
-    auto status =
-        RegOpenKeyExW(root_key, sub_key, options, sam_desired, &handle);
-    if (status == ERROR_SUCCESS) {
-      m_handle = handle;
-    }
-  }
-
-  explicit reg_key(HKEY root_key, const std::wstring &sub_key, DWORD options,
-                   REGSAM sam_desired)
-      : reg_key(root_key, sub_key.c_str(), options, sam_desired) {}
-
-  virtual ~reg_key() {
-    if (m_handle) {
-      RegCloseKey(m_handle);
-      m_handle = nullptr;
-    }
-  }
-
-  reg_key(const reg_key &other) = delete;
-  reg_key &operator=(const reg_key &other) = delete;
-  reg_key(reg_key &&other) = delete;
-  reg_key &operator=(reg_key &&other) = delete;
-
-  bool is_open() const { return !!m_handle; }
-  bool get_handle() const { return m_handle; }
-
-  std::wstring query_string(const wchar_t *name) const {
-    DWORD buf_length = 0;
-    // Get the size of the data in bytes.
-    auto status = RegQueryValueExW(m_handle, name, nullptr, nullptr, nullptr,
-                                   &buf_length);
-    if (status != ERROR_SUCCESS && status != ERROR_MORE_DATA) {
-      return std::wstring();
-    }
-    // Read the data.
-    std::wstring result(buf_length / sizeof(wchar_t), 0);
-    auto buf = reinterpret_cast<LPBYTE>(&result[0]);
-    status =
-        RegQueryValueExW(m_handle, name, nullptr, nullptr, buf, &buf_length);
-    if (status != ERROR_SUCCESS) {
-      return std::wstring();
-    }
-    // Remove trailing null-characters.
-    for (std::size_t length = result.size(); length > 0; --length) {
-      if (result[length - 1] != 0) {
-        result.resize(length);
-        break;
-      }
-    }
-    return result;
-  }
-
-private:
-  HKEY m_handle = nullptr;
-};
-
 inline bool enable_dpi_awareness() {
   auto user32 = native_library(L"user32.dll");
   if (auto fn = user32.get(user32_symbols::SetProcessDpiAwarenessContext)) {
@@ -370,6 +309,8 @@ static constexpr auto GetAvailableCoreWebView2BrowserVersionString =
         "GetAvailableCoreWebView2BrowserVersionString");
 } // namespace webview2_symbols
 #endif /* WEBVIEW_MSWEBVIEW2_EXPLICIT_LINK */
+
+#include "webview2_loader/reg_key.hpp"
 
 class loader {
 public:
